refactor(client): improved client class and models
Added decimal to manage the amount of the transaction, allowing more precise
numbers. Also improved the client class to be able to scale to multiple flavors.
Finally API models where improved to not use as many empty values when null.

+++ repositories/client/lib/src/client_base.dart
@@ -1,4 +1,5 @@
 import 'package:client/src/models/models.dart';
+import 'package:decimal/decimal.dart';
 
 /// Base class for the client.
 abstract class ClientBase {
@@ -8,7 +9,7 @@ abstract class ClientBase {
     required ExchangeType type,
     required String cryptoCurrencyId,
     required String fiatCurrencyId,
-    required double amount,
+    required Decimal amount,
     required String amountCurrencyId,
   });
